Set C bindings for network information in their own files
