+++ src/unix/jbuild
@@ -200,6 +200,9 @@
     unix_gethostname_job
     unix_gethostbyname_job
     unix_gethostbyaddr_job
+    unix_getprotoby_getservby_job
+    unix_getaddrinfo_job
+    unix_getnameinfo_job
     not_available
   ))
   (install_c_headers (
