+++ src/unix/lwt_unix_unix.h
@@ -28,7 +28,6 @@
    function, including conceptual documentation, practical considerations,
    common pitfalls, etc. */
 
-#define ARGS(args...) args
 
 #include <caml/unixsupport.h>
 #include <caml/version.h>
@@ -40,398 +39,6 @@
 #include <sys/un.h>
 #include <sys/wait.h>
 #include <unistd.h>
-
-#include "unix_recv_send_utils.h"
-
-/* Helper functions for not re-entrant functions */
-
-#include "unix_get_network_information_utils.h"
-/* +-----------------------------------------------------------------+
-   | JOBs: getprotoby{name,number}, getservby{name,port}             |
-   +-----------------------------------------------------------------+ */
-
-static value alloc_protoent(struct protoent *entry)
-{
-    value res;
-    value name = Val_unit, aliases = Val_unit;
-
-    Begin_roots2(name, aliases);
-    name = caml_copy_string(entry->p_name);
-    aliases = caml_copy_string_array((const char **)entry->p_aliases);
-    res = caml_alloc_small(3, 0);
-    Field(res, 0) = name;
-    Field(res, 1) = aliases;
-    Field(res, 2) = Val_int(entry->p_proto);
-    End_roots();
-    return res;
-}
-
-static value alloc_servent(struct servent *entry)
-{
-    value res;
-    value name = Val_unit, aliases = Val_unit, proto = Val_unit;
-
-    Begin_roots3(name, aliases, proto);
-    name = caml_copy_string(entry->s_name);
-    aliases = caml_copy_string_array((const char **)entry->s_aliases);
-    proto = caml_copy_string(entry->s_proto);
-    res = caml_alloc_small(4, 0);
-    Field(res, 0) = name;
-    Field(res, 1) = aliases;
-    Field(res, 2) = Val_int(ntohs(entry->s_port));
-    Field(res, 3) = proto;
-    End_roots();
-    return res;
-}
-
-#if defined(HAVE_NETDB_REENTRANT)
-
-#define JOB_GET_ENTRY2(INIT, FUNC, TYPE, ARGS_VAL, ARGS_DECL, ARGS_CALL)     \
-    struct job_##FUNC {                                                      \
-        struct lwt_unix_job job;                                             \
-        struct TYPE entry;                                                   \
-        struct TYPE *ptr;                                                    \
-        char *buffer;                                                        \
-        ARGS_DECL;                                                           \
-    };                                                                       \
-                                                                             \
-    static void worker_##FUNC(struct job_##FUNC *job)                        \
-    {                                                                        \
-        size_t size = 1024;                                                  \
-        for (;;) {                                                           \
-            job->buffer = (char *)lwt_unix_malloc(size);                     \
-                                                                             \
-            int result = FUNC##_r(ARGS_CALL, &job->entry, job->buffer, size, \
-                                  &job->ptr);                                \
-                                                                             \
-            switch (result) {                                                \
-                case 0:                                                      \
-                    return;                                                  \
-                case ERANGE:                                                 \
-                    free(job->buffer);                                       \
-                    size += 1024;                                            \
-                    break;                                                   \
-                case ENOENT:                                                 \
-                default:                                                     \
-                    job->ptr = NULL;                                         \
-                    return;                                                  \
-            }                                                                \
-        }                                                                    \
-    }                                                                        \
-                                                                             \
-    static value result_##FUNC(struct job_##FUNC *job)                       \
-    {                                                                        \
-        if (job->ptr == NULL) {                                              \
-            free(job->buffer);                                               \
-            lwt_unix_free_job(&job->job);                                    \
-            caml_raise_not_found();                                          \
-        } else {                                                             \
-            value res = alloc_##TYPE(&job->entry);                           \
-            free(job->buffer);                                               \
-            lwt_unix_free_job(&job->job);                                    \
-            return res;                                                      \
-        }                                                                    \
-    }                                                                        \
-                                                                             \
-    CAMLprim value lwt_unix_##FUNC##_job(ARGS_VAL)                           \
-    {                                                                        \
-        INIT;                                                                \
-        return lwt_unix_alloc_job(&(job->job));                              \
-    }
-
-#else /* defined(HAVE_NETDB_REENTRANT) */
-
-static struct servent *servent_dup(const struct servent *serv)
-{
-    struct servent *s;
-    if (!serv) {
-        return NULL;
-    }
-    s = malloc(sizeof *s);
-    if (s == NULL) {
-        goto nomem1;
-    }
-    s->s_name = s_strdup(serv->s_name);
-    if (s->s_name == NULL) {
-        goto nomem2;
-    }
-    s->s_proto = s_strdup(serv->s_proto);
-    if (s->s_proto == NULL) {
-        goto nomem3;
-    }
-    s->s_aliases = c_copy_string_array(serv->s_aliases);
-    if (s->s_aliases == NULL && serv->s_aliases != NULL) {
-        goto nomem4;
-    }
-    s->s_port = serv->s_port;
-    return s;
-nomem4:
-    free(s->s_proto);
-nomem3:
-    free(s->s_name);
-nomem2:
-    free(s);
-nomem1:
-    return NULL;
-}
-
-static void servent_free(struct servent *s)
-{
-    if (!s) {
-        return;
-    }
-    free(s->s_proto);
-    free(s->s_name);
-    c_free_string_array(s->s_aliases);
-    free(s);
-}
-
-static struct protoent *protoent_dup(const struct protoent *proto)
-{
-    if (!proto) {
-        return NULL;
-    }
-    struct protoent *p = malloc(sizeof *p);
-    if (p == NULL) {
-        return NULL;
-    }
-    p->p_name = s_strdup(proto->p_name);
-    if (p->p_name == NULL) {
-        goto nomem1;
-    }
-    p->p_aliases = c_copy_string_array(proto->p_aliases);
-    if (p->p_aliases == NULL && proto->p_aliases != NULL) {
-        goto nomem2;
-    }
-    p->p_proto = proto->p_proto;
-    return p;
-nomem2:
-    free(p->p_name);
-nomem1:
-    free(p);
-    return NULL;
-}
-
-static void protoent_free(struct protoent *p)
-{
-    if (p) {
-        free(p->p_name);
-        c_free_string_array(p->p_aliases);
-        free(p);
-    }
-}
-
-#define JOB_GET_ENTRY2(INIT, FUNC, TYPE, ARGS_VAL, ARGS_DECL, ARGS_CALL) \
-    struct job_##FUNC {                                                  \
-        struct lwt_unix_job job;                                         \
-        struct TYPE *entry;                                              \
-        ARGS_DECL;                                                       \
-    };                                                                   \
-                                                                         \
-    static void worker_##FUNC(struct job_##FUNC *job)                    \
-    {                                                                    \
-        job->entry = FUNC(ARGS_CALL);                                    \
-        if (job->entry) {                                                \
-            job->entry = TYPE##_dup(job->entry);                         \
-            if (!job->entry) {                                           \
-            }                                                            \
-        }                                                                \
-    }                                                                    \
-                                                                         \
-    static value result_##FUNC(struct job_##FUNC *job)                   \
-    {                                                                    \
-        if (job->entry == NULL) {                                        \
-            lwt_unix_free_job(&job->job);                                \
-            caml_raise_not_found();                                      \
-        } else {                                                         \
-            value res = alloc_##TYPE(job->entry);                        \
-            TYPE##_free(job->entry);                                     \
-            lwt_unix_free_job(&job->job);                                \
-            return res;                                                  \
-        }                                                                \
-    }                                                                    \
-                                                                         \
-    CAMLprim value lwt_unix_##FUNC##_job(ARGS_VAL)                       \
-    {                                                                    \
-        INIT;                                                            \
-        return lwt_unix_alloc_job(&(job->job));                          \
-    }
-
-#endif /* defined(HAVE_NETDB_REENTRANT) */
-
-JOB_GET_ENTRY2(LWT_UNIX_INIT_JOB_STRING(job, getprotobyname, 0, name),
-               getprotobyname, protoent, value name, char *name;
-               char data[], job->name)
-JOB_GET_ENTRY2(LWT_UNIX_INIT_JOB(job, getprotobynumber, 0);
-               job->num = Int_val(num), getprotobynumber, protoent, value num,
-               int num, job->num)
-JOB_GET_ENTRY2(LWT_UNIX_INIT_JOB_STRING2(job, getservbyname, 0, name, proto),
-               getservbyname, servent, ARGS(value name, value proto),
-               char *name;
-               char *proto; char data[], ARGS(job->name, job->proto))
-JOB_GET_ENTRY2(LWT_UNIX_INIT_JOB_STRING(job, getservbyport, 0, proto);
-               job->port = htons(Int_val(port)), getservbyport, servent,
-               ARGS(value port, value proto), int port;
-               char *proto; char data[], ARGS(job->port, job->proto))
-
-/* +-----------------------------------------------------------------+
-   | JOB: getaddrinfo                                                |
-   +-----------------------------------------------------------------+ */
-
-struct job_getaddrinfo {
-    struct lwt_unix_job job;
-    char *node;
-    char *service;
-    struct addrinfo hints;
-    struct addrinfo *info;
-    int result;
-    char data[];
-};
-
-static value cst_to_constr(int n, int *tbl, int size, int deflt)
-{
-    int i;
-    for (i = 0; i < size; i++)
-        if (n == tbl[i]) return Val_int(i);
-    return Val_int(deflt);
-}
-// TODO remember to include unix_recv_send_utils.h for socket_domain_table
-static value convert_addrinfo(struct addrinfo *a)
-{
-    CAMLparam0();
-    CAMLlocal3(vres, vaddr, vcanonname);
-    union sock_addr_union sa;
-    socklen_t len;
-
-    len = a->ai_addrlen;
-    if (len > sizeof(sa)) len = sizeof(sa);
-    memcpy(&sa.s_gen, a->ai_addr, len);
-    vaddr = alloc_sockaddr(&sa, len, -1);
-    vcanonname =
-        caml_copy_string(a->ai_canonname == NULL ? "" : a->ai_canonname);
-    vres = caml_alloc_small(5, 0);
-    Field(vres, 0) = cst_to_constr(a->ai_family, socket_domain_table, 3, 0);
-    Field(vres, 1) = cst_to_constr(a->ai_socktype, socket_type_table, 4, 0);
-    Field(vres, 2) = Val_int(a->ai_protocol);
-    Field(vres, 3) = vaddr;
-    Field(vres, 4) = vcanonname;
-    CAMLreturn(vres);
-}
-
-static void worker_getaddrinfo(struct job_getaddrinfo *job)
-{
-    job->result = getaddrinfo(job->node[0] ? job->node : NULL,
-                              job->service[0] ? job->service : NULL,
-                              &job->hints, &job->info);
-}
-
-static value result_getaddrinfo(struct job_getaddrinfo *job)
-{
-    CAMLparam0();
-    CAMLlocal3(vres, e, v);
-    vres = Val_int(0);
-    if (job->result == 0) {
-        struct addrinfo *r;
-        for (r = job->info; r; r = r->ai_next) {
-            e = convert_addrinfo(r);
-            v = caml_alloc_small(2, 0);
-            Field(v, 0) = e;
-            Field(v, 1) = vres;
-            vres = v;
-        }
-    }
-    if (job->info != NULL) freeaddrinfo(job->info);
-    lwt_unix_free_job(&job->job);
-    CAMLreturn(vres);
-}
-// TODO remember to include unix_recv_send_utils.h for socket_domain_table
-
-CAMLprim value lwt_unix_getaddrinfo_job(value node, value service, value hints)
-{
-    LWT_UNIX_INIT_JOB_STRING2(job, getaddrinfo, 0, node, service);
-    job->info = NULL;
-    memset(&job->hints, 0, sizeof(struct addrinfo));
-    job->hints.ai_family = PF_UNSPEC;
-    for (/*nothing*/; Is_block(hints); hints = Field(hints, 1)) {
-        value v = Field(hints, 0);
-        if (Is_block(v)) switch (Tag_val(v)) {
-                case 0: /* AI_FAMILY of socket_domain */
-                    job->hints.ai_family =
-                        socket_domain_table[Int_val(Field(v, 0))];
-                    break;
-                case 1: /* AI_SOCKTYPE of socket_type */
-                    job->hints.ai_socktype =
-                        socket_type_table[Int_val(Field(v, 0))];
-                    break;
-                case 2: /* AI_PROTOCOL of int */
-                    job->hints.ai_protocol = Int_val(Field(v, 0));
-                    break;
-            }
-        else
-            switch (Int_val(v)) {
-                case 0: /* AI_NUMERICHOST */
-                    job->hints.ai_flags |= AI_NUMERICHOST;
-                    break;
-                case 1: /* AI_CANONNAME */
-                    job->hints.ai_flags |= AI_CANONNAME;
-                    break;
-                case 2: /* AI_PASSIVE */
-                    job->hints.ai_flags |= AI_PASSIVE;
-                    break;
-            }
-    }
-    return lwt_unix_alloc_job(&job->job);
-}
-
-/* +-----------------------------------------------------------------+
-   | JOB: getnameinfo                                                |
-   +-----------------------------------------------------------------+ */
-
-struct job_getnameinfo {
-    struct lwt_unix_job job;
-    union sock_addr_union addr;
-    socklen_t addr_len;
-    int opts;
-    char host[4096];
-    char serv[1024];
-    int result;
-};
-
-static int getnameinfo_flag_table[] = {NI_NOFQDN, NI_NUMERICHOST, NI_NAMEREQD,
-                                       NI_NUMERICSERV, NI_DGRAM};
-
-static void worker_getnameinfo(struct job_getnameinfo *job)
-{
-    job->result = getnameinfo((const struct sockaddr *)&job->addr.s_gen,
-                              job->addr_len, job->host, sizeof(job->host),
-                              job->serv, sizeof(job->serv), job->opts);
-}
-
-static value result_getnameinfo(struct job_getnameinfo *job)
-{
-    CAMLparam0();
-    CAMLlocal3(vres, vhost, vserv);
-    if (job->result) {
-        lwt_unix_free_job(&job->job);
-        caml_raise_not_found();
-    } else {
-        vhost = caml_copy_string(job->host);
-        vserv = caml_copy_string(job->serv);
-        vres = caml_alloc_small(2, 0);
-        Field(vres, 0) = vhost;
-        Field(vres, 1) = vserv;
-        lwt_unix_free_job(&job->job);
-        CAMLreturn(vres);
-    }
-}
-
-CAMLprim value lwt_unix_getnameinfo_job(value sockaddr, value opts)
-{
-    LWT_UNIX_INIT_JOB(job, getnameinfo, 0);
-    get_sockaddr(sockaddr, &job->addr, &job->addr_len);
-    job->opts = caml_convert_flag_list(opts, getnameinfo_flag_table);
-    return lwt_unix_alloc_job(&job->job);
-}
 
 /* bind */
