+++ src/unix/unix_c/unix_get_network_information_utils.h
@@ -30,6 +30,7 @@
  * - unix_gethostname_job.c
  * - unix_gethostbyname_job.c
  * - unix_gethostbyaddr_job.c
+ * - unix_getprotoby_getservby_job.c
  */
 #if !defined(LWT_ON_WINDOWS)
