+++ src/unix/windows_c/not_available.c
@@ -84,4 +84,10 @@ LWT_NOT_AVAILABLE1(unix_getgrgid_job)
 LWT_NOT_AVAILABLE1(unix_gethostname_job)
 LWT_NOT_AVAILABLE1(unix_gethostbyname_job)
 LWT_NOT_AVAILABLE1(unix_gethostbyaddr_job)
+LWT_NOT_AVAILABLE1(unix_getprotobyname_job)
+LWT_NOT_AVAILABLE1(unix_getprotobynumber_job)
+LWT_NOT_AVAILABLE2(unix_getservbyname_job)
+LWT_NOT_AVAILABLE2(unix_getservbyport_job)
+LWT_NOT_AVAILABLE3(unix_getaddrinfo_job)
+LWT_NOT_AVAILABLE2(unix_getnameinfo_job)
 #endif
